refactor: interop test
